added some plotting code to separate file
